refactor(0026): Phase 5c2 — extract Auth API binary
ADR-0026 Phase 5c2.

Moves the player + agent-token surface from Satisfactory.Presentation.Api
to the Erp.Presentation.Api.Auth binary:

- GET    /players/current
- POST   /players/{id}/agent-tokens
- GET    /players/{id}/agent-tokens
- DELETE /players/{id}/agent-tokens/{tokenId}
- GET    /api/me

DevPlayerBootstrap moves Common → Auth (it bootstraps the Player
aggregate, which Auth API now owns). The MintAgentTokenRequest +
AgentTokenView records also moved (defined locally in the Auth binary).

AppHost orchestrates Auth as a real binary now: WithHttpHealthCheck("/health")
back on, Sat Web .WithReference(authApi).WaitFor(authApi) so service
discovery wires the Web → Auth route.

Web side: new AuthApiClient typed HttpClient (base
https+http://auth-api) owns the player + agent-token methods. The 3
methods moved out of PlannerApiClient. MyAgents.razor and
MintAgentTokenDialog.razor now inject both clients (Api for planner,
Auth for player ops).

Token validation in game APIs still hits the same DB via the shared
AgentTokenAuthenticator — JWT-via-HMAC migration is phase 5c3.

Test changes:
- AgentApiFactory.MintTokenAsync + LogsApiFactory.MintTokenAsync now
  bypass HTTP and use DI to call IAgentTokenRepository directly. Lets
  the existing Sat-API-targeting fixtures keep producing tokens without
  reaching for the moved /players/* endpoint.
- PlayerTokenEndpointsTests (5 tests) skipped with explanation pointing
  at a follow-up issue. Those tests need an AuthApiFactory rebuild
  against Auth API's Program — separate PR.

Build green; ApiService.Tests 34/34 passing (5 skipped); other suites
unchanged (pre-existing PlayerScopedCatalogProvider failure on main
not touched).
