+++ src/Presentation/Api/Satisfactory.Presentation.Api/Program.cs
@@ -51,7 +51,9 @@ builder.Services.Configure<AgentTokenAuthenticatorOptions>(
     builder.Configuration.GetSection(AgentTokenAuthenticatorOptions.SectionName));
 builder.Services.AddMemoryCache();
 builder.Services.AddSingleton<IAgentTokenAuthenticator, AgentTokenAuthenticator>();
-builder.Services.AddHostedService<DevPlayerBootstrap>();
+// DevPlayerBootstrap moved to the Auth API in phase 5c2 — the Player + AgentToken
+// aggregate is owned there. Sat still validates tokens by hash via the shared
+// authenticator until 5c3 swaps token verification to JWT-via-HMAC.
 
 // Current-player accessor (ADR-0025 §2). v2 returns Auth:DevPlayerId; the
 // future authenticated adapter swaps this registration for an HttpContext-
@@ -958,153 +960,10 @@ app.MapPost("/api/agent/catalogue/satisfactory", async (
     });
 });
 
-// ---- Player + agent-token management (ADR-0025 §2, §9) -------------------
-//
-//   GET    /players/current                     — resolves the dev player
-//                                                 (Auth:DevPlayerId) for the
-//                                                 Web UI's "scope" context
-//   POST   /players/{id}/agent-tokens           — mint, plaintext shown once
-//   GET    /players/{id}/agent-tokens           — list (no plaintext)
-//   DELETE /players/{id}/agent-tokens/{tokenId} — revoke
-//   GET    /me                                  — who-am-I for the agent's
-//                                                 pairing validation call
-//
-// Note: the mint/list/revoke endpoints have no caller-auth in v2 — they're
-// intended to be reached only by the Web UI, which itself runs the
-// "single-user-shaped on purpose" dev-player flow until login lands.
-// Production deployment must hide them behind the homelab's Web UI gate.
-// ---------------------------------------------------------------------------
-
-app.MapGet("/players/current", async (
-    Microsoft.Extensions.Options.IOptions<AuthOptions> authOptions,
-    IPlayerRepository players,
-    CancellationToken ct) =>
-{
-    var playerId = new PlayerId(authOptions.Value.DevPlayerId);
-    var player = await players.GetAsync(playerId, ct).ConfigureAwait(false);
-    if (player is null)
-    {
-        // DevPlayerBootstrap should have seeded by now; missing means the
-        // deployment has a misconfigured DevPlayerId or the bootstrap
-        // failed. 503 is more honest than 404 — caller should retry.
-        return Results.Json(
-            new { error = "Current player not yet provisioned. Check Auth:DevPlayerId and startup logs." },
-            statusCode: 503);
-    }
-
-    return Results.Ok(new
-    {
-        playerId = player.Id.Value,
-        displayName = player.DisplayName,
-        createdUtc = player.CreatedUtc,
-    });
-});
-
-app.MapPost("/players/{id:guid}/agent-tokens", async (
-    Guid id,
-    MintAgentTokenRequest request,
-    IPlayerRepository players,
-    IAgentTokenRepository tokens,
-    IAgentTokenHasher hasher,
-    TimeProvider clock,
-    CancellationToken ct) =>
-{
-    var playerId = new PlayerId(id);
-    var player = await players.GetAsync(playerId, ct).ConfigureAwait(false);
-    if (player is null) return Results.NotFound(new { error = $"Player {id} not found." });
-
-    var label = string.IsNullOrWhiteSpace(request?.Label)
-        ? $"Agent {DateTime.UtcNow:yyyy-MM-dd}"
-        : request!.Label!.Trim();
-
-    var plaintext = hasher.MintPlaintext();
-    var hash = hasher.Hash(plaintext);
-    var token = new AgentToken(
-        AgentTokenId.New(),
-        playerId,
-        label,
-        hash,
-        clock.GetUtcNow().UtcDateTime);
-    await tokens.AddAsync(token, ct).ConfigureAwait(false);
-    await tokens.SaveChangesAsync(ct).ConfigureAwait(false);
-
-    return Results.Json(new
-    {
-        id = token.Id.Value,
-        plaintext,
-        label = token.Label,
-        createdUtc = token.CreatedUtc,
-    }, statusCode: 201);
-});
-
-app.MapGet("/players/{id:guid}/agent-tokens", async (
-    Guid id,
-    IPlayerRepository players,
-    IAgentTokenRepository tokens,
-    CancellationToken ct) =>
-{
-    var playerId = new PlayerId(id);
-    var player = await players.GetAsync(playerId, ct).ConfigureAwait(false);
-    if (player is null) return Results.NotFound(new { error = $"Player {id} not found." });
-
-    var list = await tokens.ListForPlayerAsync(playerId, ct).ConfigureAwait(false);
-    return Results.Ok(list.Select(t => new AgentTokenView(
-        Id: t.Id.Value,
-        Label: t.Label,
-        CreatedUtc: t.CreatedUtc,
-        LastSeenUtc: t.LastSeenUtc,
-        RevokedUtc: t.RevokedUtc)));
-});
-
-app.MapDelete("/players/{id:guid}/agent-tokens/{tokenId:guid}", async (
-    Guid id,
-    Guid tokenId,
-    IAgentTokenRepository tokens,
-    IAgentTokenAuthenticator authenticator,
-    TimeProvider clock,
-    CancellationToken ct) =>
-{
-    var token = await tokens.GetAsync(new AgentTokenId(tokenId), ct).ConfigureAwait(false);
-    if (token is null || token.PlayerId != new PlayerId(id))
-    {
-        return Results.NotFound(new { error = $"Token {tokenId} not found for player {id}." });
-    }
-
-    token.Revoke(clock.GetUtcNow().UtcDateTime);
-    await tokens.SaveChangesAsync(ct).ConfigureAwait(false);
-    authenticator.Invalidate(token.Id);
-    return Results.NoContent();
-});
-
-// /api/me prefix per #245's routing rule — the agent reaches this through
-// the Cloudflare tunnel for pair validation, so it has to sit under /api/*.
-app.MapGet("/api/me", async (
-    HttpRequest http,
-    IAgentTokenAuthenticator authenticator,
-    IPlayerRepository players,
-    CancellationToken ct) =>
-{
-    var auth = await authenticator.AuthenticateAsync(http.Headers["X-Agent-Token"].ToString(), ct).ConfigureAwait(false);
-    if (!auth.IsAuthenticated)
-    {
-        return Results.Json(new { error = "X-Agent-Token is missing or invalid." }, statusCode: 401);
-    }
-
-    var player = await players.GetAsync(auth.PlayerId, ct).ConfigureAwait(false);
-    if (player is null)
-    {
-        // Token row exists but player was deleted — treat as 401 so the
-        // agent re-pairs rather than displaying a confusing partial state.
-        return Results.Json(new { error = "Player no longer exists." }, statusCode: 401);
-    }
-
-    return Results.Ok(new
-    {
-        playerId = player.Id.Value,
-        displayName = player.DisplayName,
-        tokenId = auth.TokenId.Value,
-    });
-});
+// Player + agent-token endpoints (GET /players/current, POST/GET/DELETE
+// /players/{id}/agent-tokens, GET /api/me) moved to Erp.Presentation.Api.Auth
+// in phase 5c2 — see ADR-0026 §Presentation/Api/Auth and the auth binary's
+// Program.cs.
 
 // ---- Catalogue re-ingest control (ADR-0025 §7) ----------------------------
 //
@@ -1189,17 +1048,6 @@ public sealed record ShareTokenView(string Token, string Url, DateTime CreatedUt
 /// parse Serilog's output template, the Web component highlights levels on
 /// best-effort.</summary>
 public sealed record AgentLogsRequest(IReadOnlyList<string>? Lines, string? AgentVersion = null);
-
-/// <summary>POST /players/{id}/agent-tokens body (ADR-0025 §2).</summary>
-public sealed record MintAgentTokenRequest(string? Label);
-
-/// <summary>GET /players/{id}/agent-tokens row shape (no plaintext).</summary>
-public sealed record AgentTokenView(
-    Guid Id,
-    string Label,
-    DateTime CreatedUtc,
-    DateTime? LastSeenUtc,
-    DateTime? RevokedUtc);
 
 public partial class Program { }
