+++ src/Presentation/Web/Satisfactory.Presentation.Web/PlannerApiClient.cs
@@ -188,26 +188,7 @@ public class PlannerApiClient(HttpClient httpClient)
         return response.IsSuccessStatusCode;
     }
 
-    // ---- Agent tokens (ADR-0025 §2) ---------------------------------------
-
-    public Task<CurrentPlayerView?> GetCurrentPlayerAsync(CancellationToken ct = default) =>
-        httpClient.GetFromJsonAsync<CurrentPlayerView>("/players/current", ct);
-
-    public async Task<MintAgentTokenResponse?> MintAgentTokenAsync(Guid playerId, string? label, CancellationToken ct = default)
-    {
-        var response = await httpClient.PostAsJsonAsync($"/players/{playerId}/agent-tokens", new { label }, ct);
-        response.EnsureSuccessStatusCode();
-        return await response.Content.ReadFromJsonAsync<MintAgentTokenResponse>(ct);
-    }
-
-    public async Task<AgentTokenView[]> ListAgentTokensAsync(Guid playerId, CancellationToken ct = default) =>
-        await httpClient.GetFromJsonAsync<AgentTokenView[]>($"/players/{playerId}/agent-tokens", ct) ?? [];
-
-    public async Task<bool> RevokeAgentTokenAsync(Guid playerId, Guid tokenId, CancellationToken ct = default)
-    {
-        var response = await httpClient.DeleteAsync($"/players/{playerId}/agent-tokens/{tokenId}", ct);
-        return response.IsSuccessStatusCode;
-    }
+    // Player + agent-token methods moved to AuthApiClient in phase 5c2.
 
     // ---- Player catalogue (ADR-0025 §7) -----------------------------------
